% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all2all.R
\name{all2allControlsUI}
\alias{all2allControlsUI}
\title{all2allControlsUI}
\usage{
all2allControlsUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
returns the controls for left menu
}
\description{
Generates the controls in the left menu for an all2all plot
}
\note{
\code{all2allControlsUI}
}
\examples{
x <- all2allControlsUI("bar")
}
