% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{applyFiltersNew}
\alias{applyFiltersNew}
\title{applyFiltersNew}
\usage{
applyFiltersNew(data = NULL, input = NULL)
}
\arguments{
\item{data, }{loaded dataset}

\item{input, }{input parameters}
}
\value{
data
}
\description{
Apply filters based on foldChange cutoff and padj value.
This function adds a "Legend" column with "Up", "Down" or
"NS" values for visualization.
}
\examples{
x <- applyFiltersNew()

}
