% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barmain.R
\name{barMainPlotControlsUI}
\alias{barMainPlotControlsUI}
\title{barMainPlotControlsUI}
\usage{
barMainPlotControlsUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
returns the controls for left menu
}
\description{
Generates the controls in the left menu for a bar main plot
}
\note{
\code{barMainPlotControlsUI}
}
\examples{
x <- barMainPlotControlsUI("bar")
}
