% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcheffect.R
\name{batchMethod}
\alias{batchMethod}
\title{batchMethod}
\usage{
batchMethod(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
radio control
}
\description{
select batch effect method
}
\note{
\code{batchMethod}
}
\examples{

x <- batchMethod("batch")

}
