% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{changeClusterOrder}
\alias{changeClusterOrder}
\title{changeClusterOrder}
\usage{
changeClusterOrder(order = NULL, cld = NULL)
}
\arguments{
\item{order, }{order}

\item{cld, }{data}
}
\value{
heatmap plot area
}
\description{
change order of K-means clusters
}
\note{
\code{changeClusterOrder}
}
\examples{
x <- changeClusterOrder()
}
