% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{compareClust}
\alias{compareClust}
\title{compareClust}
\usage{
compareClust(
  dat = NULL,
  ont = "CC",
  org = "org.Hs.eg.db",
  fun = "enrichGO",
  title = "Ontology Distribution Comparison",
  pvalueCutoff = 0.01
)
}
\arguments{
\item{dat, }{data to compare clusters}

\item{ont, }{the ontology to use}

\item{org, }{the organism used}

\item{fun, }{fun}

\item{title, }{title of the comparison}

\item{pvalueCutoff, }{pvalueCutoff}
}
\value{
compared cluster
}
\description{
Compares the clustered data to be displayed within the GO Term
plots.
}
\note{
\code{compareClust}
}
\examples{
x <- compareClust()

}
