% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcheffect.R
\name{correctHarman}
\alias{correctHarman}
\title{Correct Batch Effect using Harman}
\usage{
correctHarman(input = NULL, idata = NULL, metadata = NULL)
}
\arguments{
\item{input, }{input values}

\item{idata, }{data}

\item{metadata, }{metadata}
}
\value{
data
}
\description{
Batch effect correction
}
\examples{
x <- correctHarman()
}
