% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowcountfilter.R
\name{dataLCFUI}
\alias{dataLCFUI}
\title{dataLCFUI
Creates a panel to filter low count genes and regions}
\usage{
dataLCFUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
panel
}
\description{
dataLCFUI
Creates a panel to filter low count genes and regions
}
\examples{
x <- dataLCFUI("lcf")

}
