% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all2all.R
\name{debrowserall2all}
\alias{debrowserall2all}
\title{debrowserall2all}
\usage{
debrowserall2all(input, output, session, data = NULL, cex = 2)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{data, }{a matrix that includes expression values}

\item{cex, }{the size of the dots}
}
\value{
all2all plot
}
\description{
Module for a bar plot that can be used in data prep, main plots
low count removal modules or any desired module
}
\examples{
x <- debrowserall2all()

}
