% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{debrowsercondselect}
\alias{debrowsercondselect}
\title{debrowsercondselect}
\usage{
debrowsercondselect(
  input = NULL,
  output = NULL,
  session = NULL,
  data = NULL,
  metadata = NULL
)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{data, }{count data}

\item{metadata, }{metadata}
}
\value{
main plot

panel
}
\description{
Condition selection
This is not a module. Module construction didn't used here, just use it
as functions not in a module.
}
\examples{
x <- debrowsercondselect()

}
