% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{debrowserdeanalysis}
\alias{debrowserdeanalysis}
\title{debrowserdeanalysis}
\usage{
debrowserdeanalysis(
  id,
  data = NULL,
  metadata = NULL,
  columns = NULL,
  conds = NULL,
  params = NULL
)
}
\arguments{
\item{id, }{namespace id}

\item{data, }{a matrix that includes expression values}

\item{metadata, }{metadata}

\item{columns, }{columns}

\item{conds, }{conditions}

\item{params, }{de parameters}
}
\value{
DE panel
}
\description{
Module to perform and visualize DE results.
}
\examples{
\dontrun{
x <- debrowserdeanalysis("de")
}

}
