% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{debrowserdensityplot}
\alias{debrowserdensityplot}
\title{debrowserdensityplot}
\usage{
debrowserdensityplot(input = NULL, output = NULL, session = NULL, data = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{data, }{a matrix that includes expression values}
}
\value{
density plot
}
\description{
Module for a density plot that can be used in data prep and
low count removal modules
}
\examples{
x <- debrowserdensityplot()

}
