% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{debrowserheatmap}
\alias{debrowserheatmap}
\title{debrowserheatmap}
\usage{
debrowserheatmap(input, output, session, expdata = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{expdata, }{a matrix that includes expression values}
}
\value{
heatmapply plot
}
\description{
Heatmap module to create interactive heatmaps and get selected list from
a heatmap
}
\examples{
x <- debrowserheatmap()

}
