% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowcountfilter.R
\name{debrowserlowcountfilter}
\alias{debrowserlowcountfilter}
\title{debrowserlowcountfilter}
\usage{
debrowserlowcountfilter(id, ldata = NULL)
}
\arguments{
\item{id, }{namespace id}

\item{ldata, }{loaded data}
}
\value{
main plot

panel
}
\description{
Module to filter low count genes/regions
}
\examples{
\donttest{
x <- debrowserlowcountfilter("lcf")
}

}
