% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{debrowserpcaplot}
\alias{debrowserpcaplot}
\title{debrowserpcaplot}
\usage{
debrowserpcaplot(id, pcadata = NULL, metadata = NULL)
}
\arguments{
\item{id, }{namespace id}

\item{pcadata, }{a matrix that includes expression values}

\item{metadata, }{metadata to color the plots}
}
\value{
main plot

panel
}
\description{
Module for a pca plot with its loadings
as a mainplot in debrowser
}
\examples{
\donttest{
x <- debrowserpcaplot("pca")
}

}
