% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{dendControlsUI}
\alias{dendControlsUI}
\title{dendControlsUI}
\usage{
dendControlsUI(id, dendtype = "Row")
}
\arguments{
\item{id, }{module ID}

\item{dendtype, }{Row or Col}
}
\value{
controls
}
\description{
get distance metric parameters
}
\note{
\code{dendControlsUI}
}
\examples{
x <- dendControlsUI("heatmap")
}
