% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{densityPlotControlsUI}
\alias{densityPlotControlsUI}
\title{densityPlotControlsUI}
\usage{
densityPlotControlsUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
returns the left menu
}
\description{
Generates the controls in the left menu for a densityPlot
}
\note{
\code{densityPlotControlsUI}
}
\examples{
x <- densityPlotControlsUI("density")
}
