% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{deServer}
\alias{deServer}
\title{deServer}
\usage{
deServer(input, output, session)
}
\arguments{
\item{input, }{input params from UI}

\item{output, }{output params to UI}

\item{session, }{session variable}
}
\value{
the panel for main plots;
}
\description{
Sets up shinyServer to be able to run DEBrowser interactively.
}
\note{
\code{deServer}
}
\examples{
deServer

}
