% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{distFunParamsUI}
\alias{distFunParamsUI}
\title{distFunParamsUI}
\usage{
distFunParamsUI()
}
\value{
funParams
}
\description{
get distance metric parameters
}
\note{
\code{distFunParamsUI}
}
\examples{
x <- distFunParamsUI()
}
