% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cutoffs.R
\name{fold_to_log2fc}
\alias{fold_to_log2fc}
\title{Convert fold-change cutoff to |log2FC|.}
\usage{
fold_to_log2fc(x)
}
\arguments{
\item{x}{Numeric fold-change cutoff.}
}
\value{
|log2FC| cutoff (log2(x)).
}
\description{
Convert fold-change cutoff to |log2FC|.
}
