% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{get_conditions_given_selection}
\alias{get_conditions_given_selection}
\title{get_conditions_given_selection}
\usage{
get_conditions_given_selection(metadata = NULL, selection = NULL)
}
\arguments{
\item{metadata, }{meta data table}

\item{selection, }{selection}
}
\value{
meta select box
}
\description{
Return the two set of conditions given the selection of meta select box
}
\examples{
x <- get_conditions_given_selection()
}
