% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all2all.R
\name{getAll2AllPlotUI}
\alias{getAll2AllPlotUI}
\title{getAll2AllPlotUI}
\usage{
getAll2AllPlotUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
A `shiny::uiOutput` placeholder under namespace `id` that
  the all2all module's server fills with the all-pairs scatter
  matrix plot.
}
\description{
all2all plots UI.
}
\note{
\code{getAll2AllPlotUI}
}
\examples{
x <- getAll2AllPlotUI("bar")

}
