% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barmain.R
\name{getBarMainPlot}
\alias{getBarMainPlot}
\title{getBarMainPlot}
\usage{
getBarMainPlot(
  data = NULL,
  cols = NULL,
  conds = NULL,
  cond_names = NULL,
  key = NULL,
  title = "",
  input = NULL
)
}
\arguments{
\item{data, }{count or normalized data}

\item{cols, }{cols}

\item{conds, }{conds}

\item{cond_names, }{condition names}

\item{key, }{key}

\item{title, }{title}

\item{input, }{input}
}
\description{
Makes Density plots
}
\examples{
getBarMainPlot()

}
