% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getColors}
\alias{getColors}
\title{getColors}
\usage{
getColors(domains = NULL)
}
\arguments{
\item{domains, }{domains to be colored}
}
\value{
colors
}
\description{
get colors for the domains
}
\examples{
x <- getColors()

}
