% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getColorShapeSelection}
\alias{getColorShapeSelection}
\title{getColorShapeSelection}
\usage{
getColorShapeSelection(metadata = NULL, input = NULL, session = NULL)
}
\arguments{
\item{metadata, }{metadata table}

\item{input, }{input}

\item{session, }{session}
}
\value{
Color and shape selection boxes
}
\description{
Generates the fill and shape selection boxes for PCA plots.
metadata file has to be loaded in this case
}
\examples{
x <- getColorShapeSelection()
}
