% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getComparisonPanel}
\alias{getComparisonPanel}
\title{getComparisonPanel}
\usage{
getComparisonPanel(flag = NULL)
}
\arguments{
\item{flag, }{to show the panel}
}
\description{
It shows comparison tab
}
\examples{
    x <- getComparisonPanel()
}

