% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getCondMsg}
\alias{getCondMsg}
\title{getCondMsg}
\usage{
getCondMsg(dc = NULL, input = NULL, cols = NULL, conds = NULL)
}
\arguments{
\item{dc, }{columns}

\item{input, }{selected comparison}

\item{cols, }{columns}

\item{conds, }{selected conditions}
}
\value{
return conditions
}
\description{
Generates and displays the current conditions and their samples
within the DEBrowser.
}
\note{
\code{getCondMsg}
}
\examples{
x <- getCondMsg()
}
