% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{getDataAssesmentText}
\alias{getDataAssesmentText}
\title{getDataAssesmentText
DataAssesment text}
\usage{
getDataAssesmentText()
}
\value{
help text for data assesment
}
\description{
getDataAssesmentText
DataAssesment text
}
\examples{
x <- getDataAssesmentText()

}
