% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{getDataPreparationText}
\alias{getDataPreparationText}
\title{getDataPreparationText
DataPreparation text}
\usage{
getDataPreparationText()
}
\value{
help text for data preparation
}
\description{
getDataPreparationText
DataPreparation text
}
\examples{
x <- getDataPreparationText()

}
