% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{getDEAnalysisText}
\alias{getDEAnalysisText}
\title{getDEAnalysisText
DEAnalysis text}
\usage{
getDEAnalysisText()
}
\value{
help text for DE Analysis
}
\description{
getDEAnalysisText
DEAnalysis text
}
\examples{
x <- getDEAnalysisText()

}
