% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{getDensityPlot}
\alias{getDensityPlot}
\title{getDensityPlot}
\usage{
getDensityPlot(data = NULL, input = NULL, title = "")
}
\arguments{
\item{data, }{count or normalized data}

\item{input, }{input}

\item{title, }{title}
}
\value{
A `plotly` density plot showing the per-sample distributions
  of the input count matrix (one density curve per sample).
}
\description{
Makes Density plots
}
\examples{
getDensityPlot()

}
