% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{getDensityPlotUI}
\alias{getDensityPlotUI}
\title{getDensityPlotUI}
\usage{
getDensityPlotUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
the panel for Density plots;
}
\description{
Density plot UI.
}
\note{
\code{getDensityPlotUI}
}
\examples{
x <- getDensityPlotUI("density")

}
