% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{getDEResultsUI}
\alias{getDEResultsUI}
\title{getDEResultsUI
Creates a panel to visualize DE results}
\usage{
getDEResultsUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
panel
}
\description{
getDEResultsUI
Creates a panel to visualize DE results
}
\examples{
x <- getDEResultsUI("batcheffect")

}
