% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getDomains}
\alias{getDomains}
\title{getDomains}
\usage{
getDomains(filt_data = NULL)
}
\arguments{
\item{filt_data, }{data to get the domains}
}
\value{
domains
}
\description{
Get domains for the main plots.
}
\examples{
x <- getDomains()
}
