% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{getEnrichKEGG}
\alias{getEnrichKEGG}
\title{getEnrichKEGG}
\usage{
getEnrichKEGG(genelist = NULL, pvalueCutoff = 0.01, org = "org.Hs.eg.db")
}
\arguments{
\item{genelist, }{gene list}

\item{pvalueCutoff, }{the p value cutoff}

\item{org, }{the organism used}
}
\value{
Enriched KEGG
}
\description{
Gathers the Enriched KEGG analysis data to be used within the
GO Term plots.
}
\note{
\code{getEnrichKEGG}
}
\examples{
x <- getEnrichKEGG()
}
