% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getGeneSetData}
\alias{getGeneSetData}
\title{getGeneSetData}
\usage{
getGeneSetData(data = NULL, geneset = NULL)
}
\arguments{
\item{data, }{loaded dataset}

\item{geneset, }{given gene set}
}
\value{
data
}
\description{
Gathers the specified gene set list to be used within the
DEBrowser.
}
\examples{
x <- getGeneSetData()

}
