% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopanel.R
\name{getGoPanel}
\alias{getGoPanel}
\title{getGoPanel}
\usage{
getGoPanel()
}
\value{
the panel for go term analysis;
}
\description{
Creates go term analysis panel within the shiny
display.
}
\note{
\code{getGoPanel}
}
\examples{
x <- getGoPanel()

}
