% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getGroupSelector}
\alias{getGroupSelector}
\title{getGroupSelector
Return the groups}
\usage{
getGroupSelector(metadata = NULL, input = NULL, index = 1, num = 0)
}
\arguments{
\item{metadata, }{meta data table}

\item{input, }{input params}

\item{index, }{index}

\item{num, }{num}
}
\value{
meta select box
}
\description{
getGroupSelector
Return the groups
}
\examples{
x <- getGroupSelector()
}
