% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getHeatmapUI}
\alias{getHeatmapUI}
\title{getHeatmapUI}
\usage{
getHeatmapUI(id)
}
\arguments{
\item{id, }{module ID}
}
\value{
heatmap plot area
}
\description{
Generates the left menu to be used for heatmap plots
}
\note{
\code{getHeatmapUI}
}
\examples{
x <- getHeatmapUI("heatmap")
}
