% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getHelpButton}
\alias{getHelpButton}
\title{getHelpButton
prepares a helpbutton for to go to a specific site in the documentation}
\usage{
getHelpButton(name = NULL, link = NULL)
}
\arguments{
\item{name, }{name that are going to come after info}

\item{link, }{link of the help}
}
\value{
the info button
}
\description{
getHelpButton
prepares a helpbutton for to go to a specific site in the documentation
}
\examples{
x <- getHelpButton()

}
