% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getHideLegendOnOff}
\alias{getHideLegendOnOff}
\title{getHideLegendOnOff}
\usage{
getHideLegendOnOff(id = "pca")
}
\arguments{
\item{id, }{namespace id}
}
\value{
A `shiny::radioButtons` toggling whether the PCA plot legend
  is rendered or hidden.
}
\description{
hide legend
}
\examples{
x <- getHideLegendOnOff("pca")
}
