% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{getHistogramUI}
\alias{getHistogramUI}
\title{getHistogramUI}
\usage{
getHistogramUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
the panel for PCA plots;
}
\description{
Histogram plots UI.
}
\note{
\code{getHistogramUI}
}
\examples{
x <- getHistogramUI("histogram")

}
