% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.R
\name{getIQRPlot}
\alias{getIQRPlot}
\title{getIQRPlot}
\usage{
getIQRPlot(data = NULL, input = NULL, title = "")
}
\arguments{
\item{data, }{count or normalized data}

\item{input, }{input}

\item{title, }{title}
}
\description{
Makes IQR boxplot plot
}
\examples{
getIQRPlot()

}
