% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.R
\name{getIQRPlotUI}
\alias{getIQRPlotUI}
\title{getIQRPlotUI}
\usage{
getIQRPlotUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
the panel for IQR plots;
}
\description{
IQR plot UI.
}
\note{
\code{getIQRPlotUI}
}
\examples{
x <- getIQRPlotUI("IQR")

}
