% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getKEGGModal}
\alias{getKEGGModal}
\title{getKEGGModal
prepares a modal for KEGG plots}
\usage{
getKEGGModal()
}
\value{
the info button
}
\description{
getKEGGModal
prepares a modal for KEGG plots
}
\examples{
x <- getKEGGModal()

}
