% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getLogo}
\alias{getLogo}
\title{getLogo}
\usage{
getLogo()
}
\value{
return logo
}
\description{
Generates and displays the logo to be shown within DEBrowser.
}
\note{
\code{getLogo}
}
\examples{
x <- getLogo()
}
