% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getMainPanel}
\alias{getMainPanel}
\title{getMainPanel}
\usage{
getMainPanel()
}
\value{
the panel for main plots;
}
\description{
main panel for volcano, scatter and maplot.
Barplot and box plots are in this page as well.
}
\note{
\code{getMainPanel}
}
\examples{
x <- getMainPanel()

}
