% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{getMean}
\alias{getMean}
\title{getMean}
\usage{
getMean(data = NULL, selcols = NULL)
}
\arguments{
\item{data, }{dataset}

\item{selcols, }{input cols}
}
\value{
data
}
\description{
Gathers the mean for selected condition.
}
\examples{
x <- getMean()

}
