% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getMergedComparison}
\alias{getMergedComparison}
\title{getMergedComparison}
\usage{
getMergedComparison(dc = NULL, nc = NULL, input = NULL)
}
\arguments{
\item{dc, }{data container}

\item{nc, }{the number of comparisons}

\item{input, }{input params}
}
\value{
data
}
\description{
Gathers the merged comparison data to be used within the
DEBrowser.
}
\examples{
x <- getMergedComparison()

}
