% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getMethodDetails}
\alias{getMethodDetails}
\title{get the detail boxes after DE method selected}
\usage{
getMethodDetails(num = NULL, input = NULL, ns = identity)
}
\arguments{
\item{num, }{panel that is going to be shown}

\item{input, }{user input}

\item{ns, }{Shiny module namespace function (defaults to `identity` for
non-module callers)}
}
\description{
get the detail boxes after DE method selected
}
\examples{
x <- getMethodDetails()

}
