% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getNormalizedMatrix}
\alias{getNormalizedMatrix}
\title{getNormalizedMatrix}
\usage{
getNormalizedMatrix(M = NULL, method = "TMM")
}
\arguments{
\item{M, }{numeric matrix}

\item{method, }{normalization method for edgeR. default is TMM}
}
\value{
normalized matrix
}
\description{
Normalizes the matrix passed to be used within various methods
within DEBrowser.  Requires edgeR package
}
\note{
\code{getNormalizedMatrix}
}
\examples{
x <- getNormalizedMatrix(mtcars)

}
