% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopanel.R
\name{getOrganismPathway}
\alias{getOrganismPathway}
\title{getOrganismPathway}
\usage{
getOrganismPathway(org)
}
\arguments{
\item{org, }{organism}
}
\value{
organism name for pathway
}
\description{
getOrganismPathway
}
\note{
\code{getOrganismPathway}
}
\examples{
x <- getOrganismPathway()

}
