% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getPCAPlotUI}
\alias{getPCAPlotUI}
\title{getPCAPlotUI}
\usage{
getPCAPlotUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
the panel for PCA plots;
}
\description{
PCA plots UI.
}
\note{
\code{getPCAPlotUI}
}
\examples{
x <- getPCAPlotUI("pca")

}
