% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getQCLeftMenu}
\alias{getQCLeftMenu}
\title{getQCLeftMenu}
\usage{
getQCLeftMenu(input = NULL)
}
\arguments{
\item{input, }{input values}
}
\value{
QC left menu
}
\description{
Generates the left menu to be used for QC plots within the
DEBrowser.
}
\note{
\code{getQCLeftMenu}
}
\examples{
x <- getQCLeftMenu()
}
