% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getSelHeat}
\alias{getSelHeat}
\title{getSelHeat}
\usage{
getSelHeat(expdata = NULL, input = NULL)
}
\arguments{
\item{expdata, }{selected genes}

\item{input, }{input params}
}
\value{
plot
}
\description{
heatmap selection functionality
}
\examples{
x <- getSelHeat()

}
