% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getStartPlotsMsg}
\alias{getStartPlotsMsg}
\title{getStartPlotsMsg}
\usage{
getStartPlotsMsg()
}
\value{
return start plot msg
}
\description{
Generates and displays the starting messgae to be shown once
the user has first seen the main plots page within DEBrowser.
}
\note{
\code{getStartPlotsMsg}
}
\examples{
x <- getStartPlotsMsg()
}
