% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getStartupMsg}
\alias{getStartupMsg}
\title{getStartupMsg}
\usage{
getStartupMsg()
}
\value{
return startup msg
}
\description{
Generates and displays the starting message within DEBrowser.
}
\note{
\code{getStartupMsg}
}
\examples{
x <- getStartupMsg()
}
