% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getUp}
\alias{getUp}
\title{getUp
get up regulated data}
\usage{
getUp(filt_data = NULL)
}
\arguments{
\item{filt_data, }{filt_data}
}
\value{
data
}
\description{
getUp
get up regulated data
}
\examples{
x <- getUp()

}
