% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{hideObj}
\alias{hideObj}
\title{hideObj}
\usage{
hideObj(btns = NULL)
}
\arguments{
\item{btns, }{hide group of objects with shinyjs}
}
\value{
invisible(NULL); called for the side effect of calling
  `shinyjs::hide()` on each supplied id.
}
\description{
Hides a shiny object.
}
\examples{
x <- hideObj()
}
