% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\docType{data}
\name{jsResetCode}
\alias{jsResetCode}
\title{Reload the app}
\format{An object of class \code{character} of length 1.}
\usage{
jsResetCode
}
\description{
Reload the app
}
\examples{
    x<-jsResetCode()
}
\keyword{datasets}

