% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{kmeansControlsUI}
\alias{kmeansControlsUI}
\title{kmeansControlsUI}
\usage{
kmeansControlsUI(id)
}
\arguments{
\item{id, }{module ID}
}
\value{
controls
}
\description{
get kmeans controls
}
\note{
\code{kmeansControlsUI}
}
\examples{
x <- kmeansControlsUI("heatmap")
}
