% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installpack.R
\name{loadpack}
\alias{loadpack}
\title{loadpack}
\usage{
loadpack(package_name = NULL)
}
\arguments{
\item{package_name, }{package name to be loaded}
}
\description{
load packages
}
\note{
\code{loadpack}
}
\examples{
x <- loadpack()

}
