% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cutoffs.R
\name{log2fc_to_fold}
\alias{log2fc_to_fold}
\title{Convert |log2FC| cutoff to fold-change cutoff.}
\usage{
log2fc_to_fold(x)
}
\arguments{
\item{x}{Numeric |log2FC| cutoff.}
}
\value{
Fold-change cutoff (2^x).
}
\description{
Convert |log2FC| cutoff to fold-change cutoff.
}
\examples{
log2fc_to_fold(1)   # 2
log2fc_to_fold(2)   # 4
}
