% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{mainPlotControlsUI}
\alias{mainPlotControlsUI}
\title{mainPlotControlsUI}
\usage{
mainPlotControlsUI(id)
}
\arguments{
\item{id, }{module ID}
}
\value{
mainPlotControls
}
\description{
Generates the left menu to be used for main plots
}
\note{
\code{mainPlotControlsUI}
}
\examples{
x <- mainPlotControlsUI("main")
}
