% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{mainScatterNew}
\alias{mainScatterNew}
\title{mainScatterNew}
\usage{
mainScatterNew(input = NULL, data = NULL, cond_names = NULL, source = NULL)
}
\arguments{
\item{input, }{input params}

\item{data, }{dataframe that has log2FoldChange and log10padj values}

\item{cond_names}{condition names}

\item{source, }{for event triggering to select genes}
}
\value{
scatter, volcano or MA plot
}
\description{
Creates the main scatter, volcano or MA plot to be displayed within the main
panel.
}
\examples{

x <- mainScatterNew()

}
