% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{niceKmeans}
\alias{niceKmeans}
\title{niceKmeans}
\usage{
niceKmeans(df = NULL, input = NULL, iter.max = 1000, nstart = 100)
}
\arguments{
\item{df, }{data}

\item{input, }{user inputs}

\item{iter.max, }{max iteration for kmeans clustering}

\item{nstart, }{n for kmeans clustering}
}
\value{
heatmap plot area
}
\description{
Generates hierarchially clustered K-means clusters
}
\note{
\code{niceKmeans}
}
\examples{
x <- niceKmeans()
}
