% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all2all.R
\name{panel.cor}
\alias{panel.cor}
\title{panel.cor}
\usage{
panel.cor(x, y, prefix = "rho=", cex.cor = 2, ...)
}
\arguments{
\item{x, }{numeric vector x}

\item{y, }{numeric vector y}

\item{prefix, }{prefix for the text}

\item{cex.cor, }{correlation font size}

\item{..., }{additional parameters}
}
\value{
all2all correlation plots
}
\description{
Prepares the correlations for the all2all plot.
}
\examples{
panel.cor(c(1, 2, 3), c(4, 5, 6))

}
