% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all2all.R
\name{panel.hist}
\alias{panel.hist}
\title{panel.hist}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x, }{a vector of values for which the histogram is desired}

\item{..., }{any additional params}
}
\value{
all2all histogram plots
}
\description{
Prepares the historgram for the all2all plot.
}
\examples{
panel.hist(1)

}
