% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pcaPlotControlsUI}
\alias{pcaPlotControlsUI}
\title{pcaPlotControlsUI}
\usage{
pcaPlotControlsUI(id = "pca")
}
\arguments{
\item{id, }{namespace id}
}
\value{
returns the left menu according to the selected tab;
}
\description{
Generates the PCA PLots Left menu to be displayed within the DEBrowser.
}
\note{
\code{pcaPlotControlsUI}
}
\examples{
x <- pcaPlotControlsUI("pca")
}
