% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSize.R
\name{plotSizeMarginsUI}
\alias{plotSizeMarginsUI}
\title{plotSizeMarginsUI}
\usage{
plotSizeMarginsUI(id, w = 800, h = 640, t = 20, b = 100, l = 100, r = 20)
}
\arguments{
\item{id, }{id}

\item{w, }{width}

\item{h, }{height}

\item{t, }{top margin}

\item{b, }{bottom margin}

\item{l, }{left margin}

\item{r, }{right margin}
}
\value{
size and margins controls
}
\description{
Size and margins module for plotly plots
}
\note{
\code{plotSizeMarginsUI}
}
\examples{
x <- plotSizeMarginsUI("heatmap")
}
