% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSize.R
\name{plotSizeUI}
\alias{plotSizeUI}
\title{plotSizeUI}
\usage{
plotSizeUI(id, w = 800, h = 600)
}
\arguments{
\item{id, }{id}

\item{w, }{width}

\item{h, }{height}
}
\value{
size and margins controls
}
\description{
Size module for plotly plots
}
\note{
\code{plotSizeUI}
}
\examples{
x <- plotSizeUI("heatmap")
}
