% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{prepPCADat}
\alias{prepPCADat}
\title{prepPCADat}
\usage{
prepPCADat(pca_data = NULL, metadata = NULL, input = NULL, pcx = 1, pcy = 2)
}
\arguments{
\item{pca_data, }{pca run results}

\item{metadata, }{additional meta data}

\item{input, }{input}

\item{pcx, }{x axis label}

\item{pcy, }{y axis label}
}
\value{
Color and shape from selection boxes or defaults
}
\description{
prepares pca data with metadata. If metadata doesn't exists
it puts all the sampels into a signlge group; "Conds".
}
\examples{
x <- prepPCADat()
}
