% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{push}
\alias{push}
\title{push}
\usage{
push(l, ...)
}
\arguments{
\item{l, }{that are going to push to the list}

\item{..., }{list object}
}
\value{
combined list
}
\description{
Push an object to the list.
}
\examples{
mylist <- list()
newlist <- push(1, mylist)
}
