% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{removeCols}
\alias{removeCols}
\title{removeCols}
\usage{
removeCols(cols = NULL, dat = NULL)
}
\arguments{
\item{cols, }{columns that are going to be removed from data frame}

\item{dat, }{data}
}
\value{
data
}
\description{
remove unnecessary columns
}
\examples{
x <- removeCols()

}
