% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{runDE}
\alias{runDE}
\title{runDE}
\usage{
runDE(
  data = NULL,
  metadata = NULL,
  columns = NULL,
  conds = NULL,
  params = NULL,
  return_dds = FALSE
)
}
\arguments{
\item{data, }{A matrix that includes all the expression raw counts,
rownames has to be the gene, isoform or region names/IDs}

\item{metadata, }{metadata of the matrix of expression raw counts}

\item{columns, }{is a vector that includes the columns that are going
to be analyzed. These columns has to match with the given data.}

\item{conds, }{experimental conditions. The order has to match
with the column order}

\item{params, }{all params for the DE methods}

\item{return_dds}{Logical. When TRUE the function returns a list
`list(res, dds)` for DESeq2 (so QC cards can introspect the fit) and
`list(res, dds = NULL)` for edgeR / limma. Default FALSE returns the
per-method data.frame as before.}
}
\value{
de results
}
\description{
Run DE algorithms on the selected parameters.  Output is
to be used for the interactive display.
}
\examples{
x <- runDE()

}
