% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{runHeatmap}
\alias{runHeatmap}
\title{runHeatmap}
\usage{
runHeatmap(input = NULL, session = NULL, expdata = NULL)
}
\arguments{
\item{input, }{input variables}

\item{session, }{session}

\item{expdata, }{a matrix that includes expression values}
}
\value{
heatmapply plot
}
\description{
Creates a heatmap based on the user selected parameters within shiny
}
\examples{
x <- runHeatmap()

}
