% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{runLimma}
\alias{runLimma}
\title{runLimma}
\usage{
runLimma(
  data = NULL,
  metadata = NULL,
  columns = NULL,
  conds = NULL,
  params = NULL
)
}
\arguments{
\item{data, }{A matrix that includes all the expression raw counts,
rownames has to be the gene, isoform or region names/IDs}

\item{metadata, }{metadata of the matrix of expression raw counts}

\item{columns, }{is a vector that includes the columns that are going
to be analyzed. These columns has to match with the given data.}

\item{conds, }{experimental conditions. The order has to match
with the column order}

\item{params, }{normfact: Calculate normalization factors to scale the raw
    library sizes. Values can be "TMM","RLE","upperquartile","none".
fitType, fitting method; "ls" for least squares or "robust"
    for robust regression
normBet: Normalizes expression intensities so that the
    intensities or log-ratios have similar distributions across a set of arrays.}
}
\value{
Limma results
}
\description{
Run Limma algorithm on the selected conditions.  Output is
to be used for the interactive display.
}
\examples{
x <- runLimma()

}
