% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{selectedInput}
\alias{selectedInput}
\title{selectedInput}
\usage{
selectedInput(id = NULL, num = 0, default = NULL, input = NULL)
}
\arguments{
\item{id, }{input id}

\item{num, }{panel that is going to be shown}

\item{default, }{default text}

\item{input, }{input params}
}
\description{
Selects user input conditions to run in DESeq.
}
\examples{
x <- selectedInput()

}
