% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{sepRadio}
\alias{sepRadio}
\title{sepRadio}
\usage{
sepRadio(id, name)
}
\arguments{
\item{id, }{module id}

\item{name, }{name}
}
\value{
radio control
}
\description{
Radio button for separators
}
\note{
\code{sepRadio}
}
\examples{

x <- sepRadio("meta", "metadata")

}
