% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{setBatch}
\alias{setBatch}
\title{setBatch
to skip batch effect correction batch variable set with the filter results}
\usage{
setBatch(fd = NULL)
}
\arguments{
\item{fd, }{filtered data}
}
\value{
fd data
}
\description{
setBatch
to skip batch effect correction batch variable set with the filter results
}
\examples{

x <- setBatch()

}
