% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{showObj}
\alias{showObj}
\title{showObj}
\usage{
showObj(btns = NULL)
}
\arguments{
\item{btns, }{show group of objects with shinyjs}
}
\value{
invisible(NULL); called for the side effect of calling
  `shinyjs::show()` on each supplied id.
}
\description{
Displays a shiny object.
}
\examples{
x <- showObj()
}
