% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startShiny.R
\name{startHeatmap}
\alias{startHeatmap}
\title{startHeatmap}
\usage{
startHeatmap()
}
\value{
the app
}
\description{
Starts the DEBrowser heatmap
}
\note{
\code{startHeatmap}
}
\examples{
    startHeatmap()

}
