% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{textareaInput}
\alias{textareaInput}
\title{textareaInput}
\usage{
textareaInput(id, label, value, rows = 20, cols = 35, class = "form-control")
}
\arguments{
\item{id, }{id of the control}

\item{label, }{label of the control}

\item{value, }{initial value}

\item{rows, }{the # of rows}

\item{cols, }{the # of  cols}

\item{class, }{css class}
}
\description{
Generates a text area input to be used for gene selection within
the DEBrowser.
}
\examples{
x <- textareaInput("genesetarea", "Gene Set",
  "Fgf21",
  rows = 5, cols = 35
)
}
