% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_account.R
\name{accountDropdownServer}
\alias{accountDropdownServer}
\title{Server: wires the dropdown to current_user, signup modal, signout.}
\usage{
accountDropdownServer(id)
}
\arguments{
\item{id}{Shiny module namespace id.}
}
\description{
Server: wires the dropdown to current_user, signup modal, signout.
}
