% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_account.R
\name{accountDropdownUI}
\alias{accountDropdownUI}
\title{UI: navbar account dropdown.}
\usage{
accountDropdownUI(id)
}
\arguments{
\item{id}{Shiny module namespace id.}
}
\description{
UI: navbar account dropdown.
}
