% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{applyFiltersToMergedComparison}
\alias{applyFiltersToMergedComparison}
\title{applyFiltersToMergedComparison}
\usage{
applyFiltersToMergedComparison(dc = NULL, nc = NULL, input = NULL)
}
\arguments{
\item{nc, }{the number of comparisons}

\item{input, }{input params}

\item{merged, }{merged data}
}
\value{
data
}
\description{
Gathers the merged comparison data to be used within the
DEBrowser.
}
\examples{
    x <- applyFiltersToMergedComparison()

}
