% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{checkCountData}
\alias{checkCountData}
\title{checkCountData}
\usage{
checkCountData(input = NULL)
}
\arguments{
\item{input, }{inputs}
}
\value{
error if there is a problem about the loaded data
}
\description{
Returns if there is a problem in the count data.
}
\note{
\code{checkCountData}
}
\examples{
    x <- checkCountData()

}
