% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{checkMetaData}
\alias{checkMetaData}
\title{checkMetaData}
\usage{
checkMetaData(input = NULL, counttable = NULL)
}
\arguments{
\item{input, }{input}

\item{counttable, }{counttable}
}
\value{
error if there is a problem about the loaded data
}
\description{
Returns if there is a problem in the count data.
}
\note{
\code{checkMetaData}
}
\examples{
    x <- checkMetaData()

}
