% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{compareClust}
\alias{compareClust}
\title{compareClust}
\usage{
compareClust(dat = NULL, ont = "CC", organism = "human",
  fun = "enrichGO", title = "Ontology Distribution Comparison")
}
\arguments{
\item{dat, }{data to compare clusters}

\item{ont, }{the ontology to use}

\item{organism, }{the organism used}

\item{fun, }{fun}

\item{title, }{title of the comparison}
}
\value{
compared cluster
}
\description{
compareClust
}
\note{
\code{compareClust}
}
\examples{
   x <- compareClust()
   
}

