% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_condselect.R
\name{condSelectServer}
\alias{condSelectServer}
\title{Comparison-Selection wizard server.}
\usage{
condSelectServer(id, data = NULL, metadata = NULL)
}
\arguments{
\item{id}{module namespace id (must match the id passed to `condSelectUI`).}

\item{data}{count matrix.}

\item{metadata}{sample-metadata data.frame; first column is the sample id.}
}
\value{
list with `n_comparisons`, `start_de`, `is_ready`, `comparisons_spec`.
}
\description{
Comparison-Selection wizard server.
}
