% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all2all.R
\name{debrowserall2all}
\alias{debrowserall2all}
\title{debrowserall2all}
\usage{
debrowserall2all(id, data = NULL, cex = 2)
}
\arguments{
\item{id, }{namespace id}

\item{data, }{a matrix that includes expression values}

\item{cex, }{the size of the dots}
}
\value{
all2all plot
}
\description{
Module for a bar plot that can be used in data prep, main plots
low count removal modules or any desired module
}
\examples{
\donttest{
x <- debrowserall2all("all2all")
}

}
