% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barmain.R
\name{debrowserbarmainplot}
\alias{debrowserbarmainplot}
\title{debrowserbarmainplot}
\usage{
debrowserbarmainplot(input, output, session, data = NULL, cols = NULL,
  conds = NULL, key = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{data, }{a matrix that includes expression values}

\item{cols, }{columns}

\item{conds, }{conditions}

\item{key, }{the gene or region name}
}
\value{
density plot
}
\description{
Module for a bar plot that can be used in data prep, main plots 
low count removal modules or any desired module
}
\examples{
    x <- debrowserbarmainplot()

}
