% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcheffect.R
\name{debrowserbatcheffect}
\alias{debrowserbatcheffect}
\title{debrowserbatcheffect}
\usage{
debrowserbatcheffect(id, ldata = NULL)
}
\arguments{
\item{id, }{namespace id}

\item{ldata, }{loaded data}
}
\value{
main plot

panel
}
\description{
Module to correct batch effect
}
\examples{
\dontrun{
x <- debrowserbatcheffect("batch")
}

}
