% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxmain.R
\name{debrowserboxmainplot}
\alias{debrowserboxmainplot}
\title{debrowserboxmainplot}
\usage{
debrowserboxmainplot(
  id,
  data = NULL,
  cols = NULL,
  conds = NULL,
  cond_names = NULL,
  key = NULL
)
}
\arguments{
\item{id, }{namespace id}

\item{data, }{a matrix that includes expression values}

\item{cols, }{columns}

\item{conds, }{conditions}

\item{cond_names, }{condition names}

\item{key, }{the gene or region name}
}
\value{
density plot
}
\description{
Module for a box plot that can be used in DEanalysis main part and
used heatmaps
}
\examples{
\dontrun{
x <- debrowserboxmainplot("box")
}

}
