% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxmain.R
\name{debrowserboxmainplot}
\alias{debrowserboxmainplot}
\title{debrowserboxmainplot}
\usage{
debrowserboxmainplot(input = NULL, output = NULL, session = NULL,
  data = NULL, cols = NULL, conds = NULL, key = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{data, }{a matrix that includes expression values}

\item{cols, }{columns}

\item{conds, }{conditions}

\item{key, }{the gene or region name}
}
\value{
density plot
}
\description{
Module for a box plot that can be used in DEanalysis main part and 
used heatmaps
}
\examples{
    x <- debrowserboxmainplot()

}
