% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{debrowserdataload}
\alias{debrowserdataload}
\title{debrowserdataload}
\usage{
debrowserdataload(id, nextpagebutton = NULL)
}
\arguments{
\item{id, }{namespace id}

\item{nextpagebutton, }{the name of the next page button after loading the data}
}
\value{
main plot

panel
}
\description{
Module to load count data and metadata
}
\examples{
\donttest{
x <- debrowserdataload("load")
}

}
