% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{debrowserhistogram}
\alias{debrowserhistogram}
\title{debrowserhistogram}
\usage{
debrowserhistogram(id, data = NULL)
}
\arguments{
\item{id, }{namespace id}

\item{data, }{a matrix that includes expression values}
}
\value{
histogram
}
\description{
Module for a histogram that can be used in data prep and
low count removal modules
}
\examples{
\donttest{
x <- debrowserhistogram("histogram")
}

}
