% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{debrowserhistogram}
\alias{debrowserhistogram}
\title{debrowserhistogram}
\usage{
debrowserhistogram(input = NULL, output = NULL, session = NULL, data = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{data, }{a matrix that includes expression values}
}
\value{
histogram
}
\description{
Module for a histogram that can be used in data prep and
low count removal modules
}
\examples{
x <- debrowserhistogram()

}
