% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.R
\name{debrowserIQRplot}
\alias{debrowserIQRplot}
\title{debrowserIQRplot}
\usage{
debrowserIQRplot(id, data = NULL)
}
\arguments{
\item{id, }{namespace id}

\item{data, }{a matrix that includes expression values}
}
\value{
IQR
}
\description{
Module for an IQR plot that can be used in data prep and
low count removal modules
}
\examples{
\dontrun{
x <- debrowserIQRplot("iqr")
}

}
