% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowcountfilter.R
\name{debrowserlowcountfilter}
\alias{debrowserlowcountfilter}
\title{debrowserlowcountfilter}
\usage{
debrowserlowcountfilter(
  input = NULL,
  output = NULL,
  session = NULL,
  ldata = NULL
)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{ldata, }{loaded data}
}
\value{
main plot

panel
}
\description{
Module to filter low count genes/regions
}
\examples{
x <- debrowserlowcountfilter()

}
