% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{debrowsermainplot}
\alias{debrowsermainplot}
\title{debrowsermainplot}
\usage{
debrowsermainplot(id, data = NULL, cond_names = NULL)
}
\arguments{
\item{data, }{a matrix that includes expression values}

\item{cond_names, }{condition names}

\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}
}
\value{
main plot

panel
}
\description{
Module for a scatter, volcano and ma plots that are going to be used
as a mainplot in debrowser
}
\examples{
x <- debrowsermainplot()

}
