% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{debrowserpcaplot}
\alias{debrowserpcaplot}
\title{debrowserpcaplot}
\usage{
debrowserpcaplot(id, pcadata = NULL, metadata = NULL)
}
\arguments{
\item{pcadata, }{a matrix that includes expression values}

\item{metadata, }{metadata to color the plots}

\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}
}
\value{
main plot

panel
}
\description{
Module for a pca plot with its loadings
as a mainplot in debrowser
}
\examples{
x <- debrowserpcaplot()

}
