% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{deUI}
\alias{deUI}
\title{deUI}
\usage{
deUI()
}
\value{
the page tagList for DEBrowser
}
\description{
Creates a shinyUI to be able to run DEBrowser interactively.
B1 shell: bslib::page_navbar with 5 nav panels (Data Prep / Main Plots /
QC Plots / GO Term / Tables), Slate + OK-blue theme, light/dark toggle.
}
\note{
\code{deUI}
}
\examples{
x <- deUI()

}
