% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{drawKEGG}
\alias{drawKEGG}
\title{drawKEGG}
\usage{
drawKEGG(input = NULL, dat = NULL, pid = NULL)
}
\arguments{
\item{input, }{input}

\item{dat, }{expression matrix}

\item{pid, }{pathway id}
}
\value{
enriched DO
}
\description{
draw KEGG patwhay with expression values
}
\note{
\code{drawKEGG}
}
\examples{
    x <- drawKEGG()
}
