% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_enrichment_gmt.R
\name{enrichmentGmtUI}
\alias{enrichmentGmtUI}
\title{UI for the gene-set source picker (Enrichment-tab sidebar).}
\usage{
enrichmentGmtUI(id)
}
\arguments{
\item{id}{Module ID.}
}
\value{
Shiny tagList for inclusion in a `bslib::accordion_panel`.
}
\description{
UI for the gene-set source picker (Enrichment-tab sidebar).
}
