% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_enrichment.R
\name{enrichmentUI}
\alias{enrichmentUI}
\title{UI for the Enrichment tab.}
\usage{
enrichmentUI(id)
}
\arguments{
\item{id}{Module ID.}
}
\value{
A `bslib::layout_sidebar` tagList.
}
\description{
Lives next to the existing GO Term tab. E2 will fold GO Term into here
as the ORA mode; E11 adds the "Compare methods" mode.
}
