% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxmain.R
\name{getBoxMainPlot}
\alias{getBoxMainPlot}
\title{getBoxMainPlot}
\usage{
getBoxMainPlot(data = NULL, cols = NULL, conds = NULL, key = NULL,
  title = "", input = NULL)
}
\arguments{
\item{data, }{count or normalized data}

\item{cols, }{cols}

\item{conds, }{conds}

\item{key, }{key}

\item{title, }{title}

\item{input, }{input}
}
\description{
Makes Density plots
}
\examples{
    getBoxMainPlot()

}
