% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getCovariateDetails}
\alias{getCovariateDetails}
\title{getCovariateDetails}
\usage{
getCovariateDetails(num = NULL, input = NULL, metadata = NULL)
}
\arguments{
\item{num, }{panel that is going to be shown}

\item{input, }{user input}

\item{metadata, }{metadata}
}
\description{
get the covariate detail box after DE method selected
}
\examples{
x <- getCovariateDetails()

}
