% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{getEnrichGO}
\alias{getEnrichGO}
\title{getEnrichGO}
\usage{
getEnrichGO(genelist, pvalueCutoff = 0.01, org = "human", ont = "CC")
}
\arguments{
\item{genelist, }{gene list}

\item{pvalueCutoff, }{p value cutoff}

\item{org, }{the organism used}

\item{ont, }{the ontology used}
}
\value{
Enriched GO
}
\description{
getEnrichGO
}
\note{
\code{getEnrichGO}
}
\examples{
   genelist<-getGeneList(c('OCLN', 'ABCC2'))
   x <- getEnrichGO(genelist, 0.01, NULL, "CC")
   
}

