% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{getGSEA}
\alias{getGSEA}
\title{getGSEA}
\usage{
getGSEA(
  dataset = NULL,
  pvalueCutoff = 0.01,
  org = "org.Hs.eg.db",
  sortfield = "log2FoldChange"
)
}
\arguments{
\item{dataset, }{dataset}

\item{pvalueCutoff, }{the p value cutoff}

\item{org, }{the organism used}

\item{sortfield, }{sort field for GSEA}
}
\value{
GSEA
}
\description{
Gathers the Enriched KEGG analysis data to be used within the
GO Term plots.
}
\note{
\code{getGSEA}
}
\examples{
    x <- getGSEA()
}
