% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getIntHeatmap}
\alias{getIntHeatmap}
\title{getIntHeatmap}
\usage{
getIntHeatmap(heatdat = NULL, count = NULL, init_data = NULL,
  lbheat = NULL, randstr = NULL)
}
\arguments{
\item{heatData, }{heatData}

\item{inputQCPlot, }{input params for QC}

\item{linked}{brush object for selection}
}
\value{
plot
}
\description{
getIntHeatmap
}
\examples{
    load(system.file("extdata", "demo", "demodata.Rda", package = "debrowser"))
    varied <- getMostVariedList(demodata, c(2:7), 50, 100)[,c(2:7)]
    heatdat<- runHeatmap(varied)
    lbheat <- linked_brush(keys = rownames(t(heatdat$carpet)))
    getIntHeatmap(t(heatdat$carpet), lbheat)

}

