% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getIntHeatmap}
\alias{getIntHeatmap}
\title{getIntHeatmap}
\usage{
getIntHeatmap(heatdat = NULL, count = NULL, init_data = NULL,
  lbheat = NULL)
}
\arguments{
\item{heatdat, }{heatData}

\item{count, }{count}

\item{init_data, }{initial data}

\item{lbheat, }{linked brush object}

\item{linked}{brush object for selection}
}
\value{
plot
}
\description{
getIntHeatmap
}
\examples{
    getIntHeatmap()

}

