% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R, R/uifuncs.R
\name{getKEGGModal}
\alias{getKEGGModal}
\title{getKEGGModal
prepares a modal for KEGG plots}
\usage{
getKEGGModal()

getKEGGModal()
}
\value{
the info button

the info button
}
\description{
getKEGGModal
prepares a modal for KEGG plots

getKEGGModal
prepares a helpbutton for to go to a specific site in the documentation
}
\examples{
x <- getKEGGModal()

x <- getKEGGModal()

}
