% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R, R/addSection.R
\name{getLeftMenu}
\alias{getLeftMenu}
\title{Left menu for QC plots}
\usage{
getLeftMenu()

getLeftMenu()
}
\value{
returns the left menu according to the selected tab;

returns the left menu according to the selected tab;
}
\description{
Left menu for QC plots

Left menu for additional plots
}
\note{
\code{getLeftMenu}

\code{getLeftMenu}
}
\examples{
    x <- getLeftMenu()

 
   x <- getLeftMenu()

}

