% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getLegendSelect}
\alias{getLegendSelect}
\title{getLegendSelect}
\usage{
getLegendSelect(id = "pca")
}
\arguments{
\item{id, }{namespace id}
}
\value{
A `shiny::selectInput` letting the user pick whether the PCA
  plot's legend tracks the color or shape mapping.
}
\description{
select legend
}
\note{
\code{getLegendSelect}
}
\examples{
x <- getLegendSelect("pca")
}
