% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getMetaSelector}
\alias{getMetaSelector}
\title{getMetaSelector}
\usage{
getMetaSelector(metadata = NULL, input = NULL, n = 0, ns = identity)
}
\arguments{
\item{metadata, }{meta data table}

\item{input, }{input params}

\item{n, }{the box number}

\item{ns, }{Shiny module namespace function (defaults to `identity` for
non-module callers)}
}
\value{
meta select box
}
\description{
Return the sample selection box using meta data table
}
\examples{
x <- getMetaSelector()
}
