% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getPCAexplained}
\alias{getPCAexplained}
\title{getPCAexplained}
\usage{
getPCAexplained(datasetInput = NULL, pca_data = NULL, input = NULL)
}
\arguments{
\item{datasetInput, }{selected data}

\item{pca_data, }{from user}

\item{input, }{input params}
}
\value{
explained plot
}
\description{
Creates a more detailed plot using the PCA results from
the selected dataset.
}
\examples{
load(system.file("extdata", "demo", "demodata.Rda", package = "debrowser"))
input <- c()
input$qcplot <- "pca"
input$col_list <- colnames(demodata[, 1:6])
dat <- getNormalizedMatrix(demodata[, 1:6])
pca_data <- run_pca(dat)
x <- getPCAexplained(dat, pca_data, input)

}
