% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getSelectedCols}
\alias{getSelectedCols}
\title{getSelectedCols}
\usage{
getSelectedCols(data = NULL, datasetInput = NULL, input = NULL)
}
\arguments{
\item{data, }{all loaded data}

\item{datasetInput, }{selected dataset}

\item{input, }{user input params}
}
\description{
gets selected columns
}
\examples{
getSelectedCols()

}
