% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getSelHeat}
\alias{getSelHeat}
\title{getSelHeat}
\usage{
getSelHeat(init_data = NULL, inputQCPlot = NULL, heatdat = NULL,
  input = NULL)
}
\arguments{
\item{init_data, }{initial data}

\item{inputQCPlot, }{count}

\item{heatdat, }{heatData}

\item{input, }{input params}
}
\value{
plot
}
\description{
heatmap selection functionality
}
\examples{
    getSelHeat()

}

