% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapUI}
\alias{heatmapUI}
\title{heatmapUI}
\usage{
heatmapUI(input, output, session)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}
}
\value{
the panel for heatmapUI;
}
\description{
Creates a shinyUI to be able to run DEBrowser interactively.
}
\note{
\code{heatmapUI}
}
\examples{
\donttest{
  shiny::shinyApp(ui = heatmapUI, server = function(input, output) {})
}

}
