% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Plot PCA results.}
\usage{
plot_pca(x = NULL, pcx = 1, pcy = 2, explained = NULL,
  metadata = NULL, color = NULL, shape = NULL, size = NULL,
  factors = NULL)
}
\arguments{
\item{x, }{dataframe with data}

\item{pcx, }{x axis label}

\item{pcy, }{y axis label}

\item{explained, }{additional axis data}

\item{metadata, }{additional data}

\item{color, }{color for plot}

\item{shape, }{shape for plot}

\item{size, }{size of the plot}

\item{factors, }{factors of the plot}
}
\value{
pca list
}
\description{
Plot PCA results.
}
\examples{
  plot_pca()
  
}

