% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSize.R
\name{plotTypeUI}
\alias{plotTypeUI}
\title{plotTypeUI}
\usage{
plotTypeUI(id)
}
\arguments{
\item{id, }{id}
}
\value{
size and margins controls
}
\description{
Plot download type
}
\note{
\code{plotTypeUI}
}
\examples{
    x <- plotTypeUI("heatmap")
}
