% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{removeExtraCols}
\alias{removeExtraCols}
\title{removeExtraCols}
\usage{
removeExtraCols(dat = NULL)
}
\arguments{
\item{dat, }{selected data}
}
\value{
`dat` with QC-extraneous columns (padj/foldChange/Legend/etc.)
  stripped, leaving only sample columns suitable for plotting.
}
\description{
remove extra columns for QC plots
}
\examples{
removeExtraCols()

}
