% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{run_pca}
\alias{run_pca}
\title{PCA analysis}
\usage{
run_pca(x = NULL, retx = TRUE, center = TRUE, scale = TRUE)
}
\arguments{
\item{x, }{dataframe with experiment data}

\item{retx, }{specifies if the data should be returned}

\item{center, }{center the PCA (Boolean)}

\item{scale, }{scale the PCA (Boolean)}
}
\value{
pca list
}
\description{
PCA analysis
}
\examples{
  x<-run_pca()
  
}

