% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{togglePanels}
\alias{togglePanels}
\title{togglePanels}
\usage{
togglePanels(num = NULL, nums = NULL, session = NULL)
}
\arguments{
\item{num, }{selected panel}

\item{nums, }{all panels}

\item{session, }{session info}
}
\value{
invisible(NULL); called for the navbar nav_show / nav_hide /
  nav_select side effects on the methodtabs nav.
}
\description{
User defined toggle to display which panels are to be shown within
DEBrowser.
}
\note{
\code{togglePanels}
}
\examples{
x <- togglePanels()
}
